Airflow/Custom_PyPackage_Airflow/requirements.txt
pandas
matplotlib
seaborn
scikit-learn
apache-airflow-providers-amazon